Refactored everything to its own component
