+++ part2/courseinfo/src/App.js
@@ -1,28 +1,4 @@
-const Header = (props) => {
-  return <h1>{props.course}</h1>
-}
-
-const Content = (props) => {
-  return (
-    <>
-      <Part part={props.parts[0]} />
-      <Part part={props.parts[1]} />
-      <Part part={props.parts[2]} />
-    </>
-  )
-}
-
-const Part = (props) => {
-  return (
-    <p>
-      {props.part.name} {props.part.exercises}
-    </p>
-  )
-}
-
-const Footer = (props) => {
-  return <p>Number of exercises {props.parts[0].exercises + props.parts[1].exercises + props.parts[2].exercises}</p>
-}
+import Course from "./components/Course";
 
 const App = () => {
   const course = {
@@ -43,14 +19,7 @@ const App = () => {
     ]
   }
 
-  return (
-    <div>
-      <Header course={course.name} />
-
-      <Content parts={course.parts} />
-      <Footer parts={course.parts} />
-    </div>
-  );
+  return <Course course={course} />
 }
 
 export default App;
