+++ part2/courseinfo/src/components/Content.js
@@ -1,0 +1,13 @@
+import Part from "./Part";
+
+const Content = (props) => {
+  return (
+    <>
+      <Part part={props.parts[0]} />
+      <Part part={props.parts[1]} />
+      <Part part={props.parts[2]} />
+    </>
+  )
+}
+
+export default Content;
