+++ part2/courseinfo/src/components/Course.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+import Header from './Header';
+import Content from './Content';
+import Footer from './Footer';
+
+export default function Course({course}) {
+  return (
+    <div>
+      <Header course={course.name} />
+      <Content parts={course.parts} />
+      <Footer parts={course.parts} />
+    </div>
+  );
+}
