+++ part2/courseinfo/src/components/Footer.js
@@ -1,0 +1,5 @@
+const Footer = (props) => {
+  return <p>Number of exercises {props.parts[0].exercises + props.parts[1].exercises + props.parts[2].exercises}</p>
+}
+
+export default Footer;
