+++ part2/courseinfo/src/components/Header.js
@@ -1,0 +1,5 @@
+const Header = (props) => {
+  return <h1>{props.course}</h1>
+}
+
+export default Header
