+++ part2/courseinfo/src/components/Part.js
@@ -1,0 +1,9 @@
+const Part = (props) => {
+  return (
+    <p>
+      {props.part.name} {props.part.exercises}
+    </p>
+  )
+}
+
+export default Part;
